% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExamCli}
\alias{ExamCli}
\title{Clinical information of Example data
From AMC}
\format{
An object of class \code{data.frame} with 90 rows and 5 columns.
}
\usage{
ExamCli
}
\description{
Clinical information of Example data
From AMC
}
\keyword{datasets}
