% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExamExp}
\alias{ExamExp}
\title{Gene expression of Example data
From AMC
row respresents gene
column represents sample}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6137 rows and 90 columns.
}
\usage{
ExamExp
}
\description{
Gene expression of Example data
From AMC
row respresents gene
column represents sample
}
\keyword{datasets}
