% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIclassifier.R
\name{GIclassifier}
\alias{GIclassifier}
\title{description >> Function for the classification of GI cancer subtypes
argument}
\usage{
GIclassifier(Expr, idType = c("SYMBOL", "ENSEMBL", "ENTREZID"), verbose = TRUE)
}
\arguments{
\item{Expr}{a dataframe with log2_scaled Gene Expression Profiles data values, samples in columns, genes in rows}

\item{idType}{the type of geneid "SYMBOL", "ENSEMBL", "ENTREZID"}

\item{verbose}{TRUE/FALSE}
}
\description{
description >> Function for the classification of GI cancer subtypes
argument
}
