% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TrainExp}
\alias{TrainExp}
\title{Gene expression of Training data for GI cancer subtype
row respresents gene
column represents GI cancer subtype}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6137 rows and 4 columns.
}
\usage{
TrainExp
}
\description{
Gene expression of Training data for GI cancer subtype
row respresents gene
column represents GI cancer subtype
}
\keyword{datasets}
